Consume scheduled replication events
Prior to this change replication-status was only consuming completed
replication events, that is, replication events that already completed,
either successfully or not.

Now, consume also scheduled replication events, so that when querying
the REST-API it is possible to have visibility on replication events
that are currently scheduled, but have yet to be completed.

Bug: Issue 14802
Change-Id: I49fb3d5cf94732e5193529336fa224a87c561d63