Request by remote name instead of remote URL
The replication-status currently accepts a remote-url as a URL segment.
This can be tedious from a client point of view because the fully qualified
replication-url might not be know plus user cannot ask for multiple urls
which are logically grouped. To mitigate this situation use remote name
instead of remote url.

Bug: Issue 14803
Change-Id: Ibd6c2a943204cacc035b74fa34e34220417274c3