Document required installation of replication plugin
The replication-status plugin relies on events defined in the
replication plugin.

Since Gerrit doesn't provide a way of defining dependencies among plugins,
if replication-status loads before replication, some classes won't be available
at runtime.

As a workaround the replication plugin can be added as a lib to
make sure the needed classes are available at startup.

Bug: Issue 16554
Change-Id: I5a8b0faa6009759249ce6277633db1c75b08bcda